rename data_nodes module to nodes
